+++ sakura_core/HokanMgr.h
@@ -34,7 +34,7 @@ public:
 	HWND DoModeless(HINSTANCE, HWND, LPARAM); // モードレスダイアログの表示
 	void Hide(void);
 	// 初期化
-	int Search(
+	size_t Search(
 		POINT*			ppoWin,
 		int				nWinHeight,
 		int				nColumnWidth,
