+++ sakura_core/doc/DocReader.h
@@ -30,9 +30,9 @@ class DocReader {
 public:
 	DocReader(const DocLineMgr& pcDocLineMgr) : pDocLineMgr(&pcDocLineMgr) { }
 
-	wchar_t* GetAllData(int* pnDataLen);	// 全行データを返す
-	const wchar_t* GetLineStr(int , size_t*);
-	const wchar_t* GetLineStrWithoutEOL(int , int*); // 2003.06.22 Moca
+	wchar_t* GetAllData(size_t* pnDataLen);	// 全行データを返す
+	const wchar_t* GetLineStr(size_t , size_t*);
+	const wchar_t* GetLineStrWithoutEOL(size_t , size_t*); // 2003.06.22 Moca
 	const wchar_t* GetFirstLinrStr(size_t*);	// 順アクセスモード：先頭行を得る
 	const wchar_t* GetNextLinrStr(size_t*);	// 順アクセスモード：次の行を得る
